Move generic functions to JSONAPIPlug
