+++ lib/jsonapi_plug/resource.ex
@@ -76,7 +76,7 @@ defmodule JSONAPIPlug.Resource do
       config :my_app, MyApp.API, host: "adifferenthost.com"
   """
 
-  alias JSONAPIPlug.{API, Document, Document.ResourceObject, Normalizer}
+  alias JSONAPIPlug.{Document, Document.ResourceObject, Normalizer}
   alias Plug.Conn
 
   @attribute_schema [
@@ -178,13 +178,6 @@ defmodule JSONAPIPlug.Resource do
   Resource data is either a resource or a list of resources
   """
   @type data :: resource() | [resource()]
-
-  @typedoc """
-  Resource meta
-
-  A free form map containing metadata to be rendered
-  """
-  @type meta :: Document.meta()
 
   @typedoc """
   Resource field name
@@ -380,7 +373,7 @@ defmodule JSONAPIPlug.Resource do
                 Document.t() | no_return()
         def render(action, assigns)
             when action in ["create.json", "index.json", "show.json", "update.json"] do
-          JSONAPIPlug.Resource.render(
+          JSONAPIPlug.render(
             __MODULE__,
             assigns[:conn],
             assigns[:data],
@@ -425,117 +418,4 @@ defmodule JSONAPIPlug.Resource do
   def field_option(field, _option, _default) do
     raise "invalid field definition: #{inspect(field)}"
   end
-
-  @doc """
-  Related Resource based on JSON:API type
-
-  Returns the resource used to handle relationships of the requested type by the passed resource.
-  """
-  @spec for_related_type(t(), ResourceObject.type()) :: t() | nil
-  def for_related_type(resource, type) do
-    Enum.find_value(resource.relationships(), fn {_relationship, options} ->
-      relationship_resource = Keyword.fetch!(options, :resource)
-
-      if relationship_resource.type() == type do
-        relationship_resource
-      else
-        nil
-      end
-    end)
-  end
-
-  @doc """
-  Render JSON:API response
-
-  Renders the JSON:API response for the specified Resource.
-  """
-  @spec render(t(), Conn.t(), data() | nil, Document.meta() | nil, options()) ::
-          Document.t() | no_return()
-  def render(
-        resource,
-        conn,
-        data \\ nil,
-        meta \\ nil,
-        options \\ []
-      ) do
-    resource
-    |> Normalizer.normalize(conn, data, meta, options)
-    |> Document.serialize()
-  end
-
-  @doc """
-  Generate relationships link
-
-  Generates the relationships link for a resource.
-  """
-  @spec url_for_relationship(t(), resource(), Conn.t() | nil, ResourceObject.type()) :: String.t()
-  def url_for_relationship(resource, data, conn, relationship_type) do
-    Enum.join([url_for(resource, data, conn), "relationships", relationship_type], "/")
-  end
-
-  @doc """
-  Generates the resource link
-
-  Generates the resource link for a resource.
-  """
-  @spec url_for(t(), data(), Conn.t() | nil) :: String.t()
-  def url_for(resource, data, conn) when is_nil(resource) or is_list(data) do
-    conn
-    |> render_uri([resource.path() || resource.type()])
-    |> to_string()
-  end
-
-  def url_for(
-        resource,
-        data,
-        %Conn{private: %{jsonapi_plug: %JSONAPIPlug{} = jsonapi_plug}} = conn
-      ) do
-    normalizer = resource.normalizer() || API.get_config(jsonapi_plug.api, [:normalizer])
-
-    conn
-    |> render_uri([
-      resource.path() || resource.type(),
-      normalizer.normalize_attribute(data, resource.id_attribute())
-    ])
-    |> to_string()
-  end
-
-  defp render_uri(%Conn{} = conn, path) do
-    %URI{
-      scheme: scheme(conn),
-      host: host(conn),
-      path: Enum.join([namespace(conn) | path], "/"),
-      port: port(conn)
-    }
-  end
-
-  defp render_uri(_conn, path), do: %URI{path: "/" <> Enum.join(path, "/")}
-
-  defp scheme(%Conn{private: %{jsonapi_plug: %JSONAPIPlug{} = jsonapi_plug}, scheme: scheme}),
-    do: to_string(API.get_config(jsonapi_plug.api, [:scheme], scheme))
-
-  defp scheme(_conn), do: nil
-
-  defp host(%Conn{private: %{jsonapi_plug: %JSONAPIPlug{} = jsonapi_plug}, host: host}),
-    do: API.get_config(jsonapi_plug.api, [:host], host)
-
-  defp host(_conn), do: nil
-
-  defp namespace(%Conn{private: %{jsonapi_plug: %JSONAPIPlug{} = jsonapi_plug}}) do
-    case API.get_config(jsonapi_plug.api, [:namespace]) do
-      nil -> ""
-      namespace -> "/" <> namespace
-    end
-  end
-
-  defp namespace(_conn), do: ""
-
-  defp port(%Conn{private: %{jsonapi_plug: %JSONAPIPlug{} = jsonapi_plug}, port: port} = conn) do
-    case API.get_config(jsonapi_plug.api, [:port], port) do
-      nil -> nil
-      port -> if port == URI.default_port(scheme(conn)), do: nil, else: port
-    end
-  end
-
-  defp port(_conn), do: nil
 end
